+++ playground/src/main.ts
@@ -2,5 +2,6 @@ import { createApp } from "vue";
 import App from "./App.vue";
 import "@seclab-dev/tokens/index.css";
 import "@seclab-dev/vue/style.css";
+import "./index.css";
 
 createApp(App).mount("#app");